fixed dtypes

+++ jukebox/jukebox/utils/audio_utils.py
@@ -159,9 +159,9 @@ def load_batches_of_embeddings(path_to_data, hps, model, use_level) -> t.Tensor:
     
     print_once(f'Shape of loaded data: {data.shape}')
     num_of_encoded_samples = data.shape[0] // encoded_sequence_length
-    num_of_embs_to_remove = (num_of_encoded_samples % hps.bs) * encoded_sequence_length
+    num_of_embs_to_remove = int((num_of_encoded_samples % hps.bs) * encoded_sequence_length)
     data = data[:data.shape[0] - num_of_embs_to_remove]
-    num_of_batches = (data.shape[0] / encoded_sequence_length) / hps.bs
+    num_of_batches = int((data.shape[0] / encoded_sequence_length) / hps.bs)
     data = data.reshape(num_of_batches, hps.bs, encoded_sequence_length)
     return data
